requirements-gtk.txt
PyGObject>=3.44.0
zeroconf>=0.115.0
aiohttp>=3.9.0
asyncio>=3.4.3
telnetlib3>=2.0.4
